feat(collect): 수집 파일을 표에 넣는 적재기 · 초기 402편 적재 (#18)
* feat(collect): 수집 파일을 표에 넣는 적재기 · 초기 402편 적재

- features/collect 에 리포지토리 6개, 파일 읽기, 저장 서비스, 관리자 API 추가
- 저장 경로는 한 곳(PostIngestService). 파일에서 넣는 입구와 앞으로 웹 수집 입구가 같이 쓴다
- 같은 글은 url 로 걸러낸다. 다시 돌려도 행이 늘지 않는다
- 기술 이름은 수집 단계와 같은 두 단계(정규화·동의어 통합)를 거친다.
  사전이 못 묶은 넷(Spring Framework→Spring, Server-Sent Events→SSE,
  Model Context Protocol→MCP, Claude Code Skill→Claude Code)을 더했다.
  이대로 두면 사용자 선택지 20종 중 넷이 글에 하나도 안 걸린다
- 본문을 소제목으로 잘라 구간을 만든다. 앞에서부터 순서대로 찾아 같은 문구가 반복돼도 안 꼬인다

엔티티 수정 2건 (표에 데이터가 없을 때 고침)
- TechTagEntity PK Integer → Long. blog_post_tag 가 Long 을 기대해 글에 태그를 못 붙이고 있었다
- BlogPostTagEntity tag_rang → tag_rank. 오타. 두면 컬럼이 둘 생긴다

기타
- build.gradle 에 실행·테스트 UTF-8. JDK 17 이 Windows 기본(MS949)으로 읽어 한글 경로가 깨졌다
- SecurityConfig 에 적재 경로만 임시 개방. ADMIN 계정이 생기면 되돌린다
- application-dev.yml 에 AI 분류 조건(모델·본문 7000자·온도 0) 기록. C3 에서 같은 값을 쓴다

확인
./gradlew clean test → BUILD SUCCESSFUL
실제 DB 적재 → 기업 8 · 기술 541 · 글 402 · 구간 4753 · 분류어 220 · 태그 899


* style: spotlessApply (구글 표준 포맷 통일)

commit 전 훅이 사람마다 다르게 깔려 있어 포맷이 안 맞는 코드가 쌓였다.
기능 변경 없음. 들여쓰기·줄바꿈·import 정리만.


---------

+++ src/main/java/com/mini8/backend/database/Tech/domain/entity/TechTagEntity.java
@@ -20,7 +20,7 @@ import lombok.NoArgsConstructor;
 public class TechTagEntity {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer tech_tag_id;
+  private Long tech_tag_id;
 
   @Column(nullable = false, unique = true, length = 50)
   private String name;
